Add adjusting on the fly
This commit adds the ability to adjust charge speed on the fly. When the car is slowcharging, you can reduce the chargespeed via HVAC controls, by setting the fan to max (7), and turning recirculation mode ON. The selected charge kW max will be displayed with both the capacity bars and SOC% if you have the newer LEAF with this option.